Rename existing toolbar to menubar
